% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{NILE}
\alias{NILE}
\title{Estimate NILE}
\usage{
NILE(
  Y,
  X,
  A,
  lambda.nile,
  intercept = TRUE,
  df = 100,
  x.new = NULL,
  p.min = 0.05,
  plot = TRUE,
  f.true = NULL,
  par.x = list(),
  par.a = list(),
  par.cv = list(num.folds = 10, optim = "optimize")
)
}
\arguments{
\item{Y, X, A, lambda.nile, intercept}{Numeric vectors. Two vectors with
\code{n} observations.}

\item{df, x.new, p.min, plot, f.true, par.cv}{Positive integer.
The number of observations used to
compute ... Set by default to \code{3.14}...}

\item{par.x, par.a}{List made of:
\itemize{
\item \code{pen.degree} this is ...
\item \code{n.order} this is ...
\item \code{num.breaks} this is ...
}}
}
\value{
Numeric --- between 0 and 1.
The causal tail coefficient between \code{v1} and \code{v2}.
}
\description{
Brief description of the function.
}
\details{
Detailed description that spans possibly several lines...
You can have lists
\enumerate{
\item ... with some equations \eqn{x^2 + y^2 = 1}...
\item or plain text
\item or reference to arguments \code{X, Y}.
}
}
