% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{predict.AR}
\alias{predict.AR}
\title{Predict an AR model}
\usage{
\method{predict}{AR}(object, xnew, ...)
}
\arguments{
\item{object}{Fitted object of class \code{AR}.}

\item{xnew}{Numeric vector with test data.}

\item{...}{Dots. Currently ignored.}
}
\value{
Numeric vector with predictions
}
\description{
... add description ...
}
